No need to cast, because meters_per_inch is already double
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/imaging/trunk@1391257 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/imaging/formats/png/PngImageParser.java
@@ -595,11 +595,11 @@ public class PngImageParser extends ImageParser implements PngConstants {
             PhysicalWidthDpi = (int) Math
                     .round(pngChunkpHYs.PixelsPerUnitXAxis
                             * meters_per_inch);
-            PhysicalWidthInch = (float) ((double) Width / ((double) pngChunkpHYs.PixelsPerUnitXAxis * meters_per_inch));
+            PhysicalWidthInch = (float) (Width / (pngChunkpHYs.PixelsPerUnitXAxis * meters_per_inch));
             PhysicalHeightDpi = (int) Math
                     .round(pngChunkpHYs.PixelsPerUnitYAxis
                             * meters_per_inch);
-            PhysicalHeightInch = (float) ((double) Height / ((double) pngChunkpHYs.PixelsPerUnitYAxis * meters_per_inch));
+            PhysicalHeightInch = (float) (Height / (pngChunkpHYs.PixelsPerUnitYAxis * meters_per_inch));
         }
 
         String FormatDetails = "Png";
